The front-end of v2.0.0

+++ src/components/AccountsPage/ManageAccountsForm.jsx
@@ -1,25 +1,29 @@
 import React from "react";
-import Input from "../Input.jsx"
+import { Input } from "../Input.jsx";
 import accountsServices from "../../services/accounts.js";
 import { useDispatch, useSelector } from "react-redux";
-import { addAccount, setAccounts, setAccountsFormIdField, setAccountsFormNameField, toggleAccountFormState } from "../../reducers/accounts.js";
+import { setId, setName, setState } from "../../reducers/accountForm.js";
+import { setAccounts, addAccount } from "../../reducers/accounts.js";
 import { removeNotification, setErrorNotification } from "../../reducers/notification.js";
-import { Cancel01Icon } from "hugeicons-react";
+import { useConfirm } from '../../customStates/useConfirm.js';
+
 import Form from "../Form.jsx";
+import transactionsServices from "../../services/transactions.js";
 const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
-  let idValue = useSelector(state => state.accounts.accountsForm.idField)
-  let nameValue = useSelector(state => state.accounts.accountsForm.nameField)
-  let formState = useSelector(state => state.accounts.accountsForm.state)
+  let id = useSelector(state => state.accountForm.id);
+  let name = useSelector(state => state.accountForm.name);
+  let formState = useSelector(state => state.accountForm.state);
+  let confirm = useConfirm();
   const dispatch = useDispatch()
   const submitHandler = async(e)=>
   {
     e.preventDefault()
-    const response = await accountsServices.createAccount(idValue, nameValue)
+    const response = await accountsServices.createAccount(id, name)
     if(response.state) 
     {
-      dispatch(addAccount({id: idValue, name: nameValue}))
-      dispatch(setAccountsFormIdField(""))
-      dispatch(setAccountsFormNameField(""))
+      dispatch(addAccount({id: id, name: name}))
+      dispatch(setId(""))
+      dispatch(setName(""))
       
     }
     else
@@ -32,24 +36,28 @@ const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
   const cancelHandler = (e)=>
   {
     e.preventDefault()
-    dispatch(toggleAccountFormState())
-    dispatch(setAccountsFormIdField(""))
-    dispatch(setAccountsFormNameField(""))
+    dispatch(setState(true))
+    dispatch(setId(""))
+    dispatch(setName(""))
   }
 
-  const removeAccountHandler = async(e)=>
+  const deleteHandler = async(e)=>
   {
-    e.preventDefault()
-    const response = await accountsServices.deleteAccount(idValue)
+    e.preventDefault();
+    const deleteConfirm = await confirm(`هل متأكد من طلبك لحذف الحساب؟`, "نعم، متأكد");
+    if(deleteConfirm === false) return;
+    const response = await accountsServices.deleteAccount(id)
     if(!response.state)
     {
-      dispatch(setErrorNotification(response.message))
+      const FOREIGN_KEY = "FOREIGN KEY constraint failed";
+      let message = response.message === FOREIGN_KEY? "ما زالت هنالك بيانات متعلقة بالموظف. يجب أن تحذف أولا." : response.message;
+      dispatch(setErrorNotification(message))
       setTimeout(()=> dispatch(removeNotification()), 3000)
       return
     }
-    dispatch(toggleAccountFormState())
-    dispatch(setAccountsFormIdField(""))
-    dispatch(setAccountsFormNameField(""))
+    dispatch(setState(true))
+    dispatch(setId(""))
+    dispatch(setName(""))
 
     const getAccountsResponse = await accountsServices.getAllAccounts()
     if(!getAccountsResponse.state) return
@@ -59,16 +67,16 @@ const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
   const editHandler = async(e)=>
   {
     e.preventDefault()
-    const response = await accountsServices.updateAccount(idValue, nameValue)
+    const response = await accountsServices.updateAccount(id, name)
     if(!response.state)
     {
       dispatch(setErrorNotification(response.message))
       setTimeout(()=> dispatch(removeNotification()), 3000)
       return
     }
-    dispatch(toggleAccountFormState())
-    dispatch(setAccountsFormIdField(""))
-    dispatch(setAccountsFormNameField(""))
+    dispatch(setState(true))
+    dispatch(setId(""))
+    dispatch(setName(""))
 
     const getAccountsResponse = await accountsServices.getAllAccounts()
     if(!getAccountsResponse.state) return
@@ -78,19 +86,19 @@ const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
   const idOnChange = (e)=>
   {
     const value = e.target.value
-    dispatch(setAccountsFormIdField(value))
+    dispatch(setId(value))
   }
 
   const nameOnChange = (e)=>
   {
     const value = e.target.value
-    dispatch(setAccountsFormNameField(value))
+    dispatch(setName(value))
   }
   
   return (
     <Form legend={legend} submitHandler={submitHandler}>
-      <Input disabled={!formState} name={'id'} label={'الكود'} type={'text'} placeholder={'أدخل كود الحساب مثل: 0000'} onChange={idOnChange} value={idValue} />
-      <Input name={'name'} label={'الإسم'} type={'text'} placeholder={'أدخل إسم الحساب مثل: أحمد'} onChange={nameOnChange} value={nameValue} />
+      <Input disabled={!formState} name={'id'} label={'الكود'} type={'text'} placeholder={'أدخل كود الحساب مثل: 0000'} onChange={idOnChange} value={id} />
+      <Input name={'name'} label={'الإسم'} type={'text'} placeholder={'أدخل إسم الحساب مثل: أحمد'} onChange={nameOnChange} value={name} />
     
       {
         formState
@@ -102,7 +110,7 @@ const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
             <button type="" className="bg-blue-500 text-white px-4 py-2 rounded-lg hover:bg-blue-600" onClick={editHandler}>
               عدل الحساب
             </button>
-            <button type="" className="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600" onClick={removeAccountHandler}>
+            <button type="" className="bg-red-500 text-white px-4 py-2 rounded-lg hover:bg-red-600" onClick={deleteHandler}>
               احذف الحساب
             </button>
             <button type="" className="bg-gray-500 text-white px-4 py-2 rounded-lg hover:bg-gray-600" onClick={cancelHandler}>
@@ -111,23 +119,6 @@ const ManageAccountsForm = ({ legend, visibale, setVisibility }) => {
           </>
         }
     </Form>
-  )
-  return (
-    <div className="fixed top-[66px] w-[98%] p-2 bg-gray-100 left-1/2 transform -translate-x-1/2 border border-gray-600">
-      <div dir="rtl">
-        <button onClick={()=> setVisibility(false)}><Cancel01Icon size={24}/></button>
-      </div>
-      <fieldset className="border border-gray-300 p-1 rounded-lg">
-        <legend className="text-lg font-semibold">{legend}</legend>
-        <form className="flex flex-col gap-2" onSubmit={submitHandler}>
-          <fieldset className="grid grid-cols-2 gap-2" dir="rtl">
-            
-          </fieldset>
-          
-
-        </form>
-      </fieldset>
-    </div>
   );
 };
